# Improve ExecuteBlocks performance.