--- a/template_tests/block_render/deep.tpl
+++ b/template_tests/block_render/deep.tpl
@@ -1,0 +1,5 @@
+{% extends "../inheritance/base3.tpl" %}
+
+{% block more_content %}
+The nested content
+{% endblock %}