--- a/template_tests/block_render/deep.tpl.out
+++ b/template_tests/block_render/deep.tpl.out
@@ -1,0 +1,4 @@
+
+The content
+
+The nested content