--- a/template_tests/inheritance/base3.tpl
+++ b/template_tests/inheritance/base3.tpl
@@ -1,0 +1,1 @@
+{% extends "inheritance3/skeleton.tpl" %}