--- a/template_tests/inheritance/inheritance3/skeleton.tpl
+++ b/template_tests/inheritance/inheritance3/skeleton.tpl
@@ -1,0 +1,5 @@
+{% extends "../inheritance2/skeleton.tpl" %}
+
+{% block content %}
+The content
+{% endblock %}